+++ news-bundle/src/Resources/contao/dca/tl_news_archive.php
@@ -269,6 +269,8 @@ class tl_news_archive extends Backend
 
 	/**
 	 * Check permissions to edit table tl_news_archive
+	 *
+	 * @throws Contao\CoreBundle\Exception\AccessDeniedException
 	 */
 	public function checkPermission()
 	{
@@ -375,8 +377,7 @@ class tl_news_archive extends Backend
 			case 'show':
 				if (!in_array(Input::get('id'), $root) || (Input::get('act') == 'delete' && !$this->User->hasAccess('delete', 'newp')))
 				{
-					$this->log('Not enough permissions to '.Input::get('act').' news archive ID "'.Input::get('id').'"', __METHOD__, TL_ERROR);
-					$this->redirect('contao/main.php?act=error');
+					throw new Contao\CoreBundle\Exception\AccessDeniedException('Not enough permissions to ' . Input::get('act') . ' news archive ID ' . Input::get('id') . '.');
 				}
 				break;
 
@@ -398,8 +399,7 @@ class tl_news_archive extends Backend
 			default:
 				if (strlen(Input::get('act')))
 				{
-					$this->log('Not enough permissions to '.Input::get('act').' news archives', __METHOD__, TL_ERROR);
-					$this->redirect('contao/main.php?act=error');
+					throw new Contao\CoreBundle\Exception\AccessDeniedException('Not enough permissions to ' . Input::get('act') . ' news archives.');
 				}
 				break;
 		}