+++ news-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.1.0-RC1 (2015-XX-XX)
 
+ * Throw an exception instead of redirecting to `/contao?act=error` (see contao/core-bundle#395).
  * Adjust the code to be compatible with PHP7 (see contao/core#8018).
  * Add the `News::generateNewsUrl()` method (see contao/core#7504).